+++ stores/authStore.js
@@ -1,9 +1,7 @@
 import { defineStore } from "pinia";
 
 export const useAuthStore = defineStore("auth", {
-  state: () => ({
-    user: null,
-  }),
+  state: () => ({user: null}),
   actions: {
     async login(formData) {
       // get api uri
@@ -42,22 +40,26 @@ export const useAuthStore = defineStore("auth", {
       navigateTo("/admin/login");
     },
     async getUser() {
-      try {
-        const config = useRuntimeConfig();
-        const apiUri = config.public.apiUri;
-        const user = await $fetch(apiUri + "/user", {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          credentials: "include",
-        });
-        this.user = user;
+      const Api = useApiStore();
+      const user = await Api.get('/user')
+      this.user = user;
+      
+      // try {
+      //   const config = useRuntimeConfig();
+      //   const apiUri = config.public.apiUri;
+      //   const user = await $fetch(apiUri + "/user", {
+      //     method: "GET",
+      //     headers: {
+      //       "Content-Type": "application/json",
+      //     },
+      //     credentials: "include",
+      //   });
+      //   this.user = user;
         
-      } catch (error) {
-        console.log('========= Error =========')
-        console.log(error)
-      }
+      // } catch (error) {
+      //   console.log('========= Error =========')
+      //   console.log(error)
+      // }
     },
   },
 });
